live clock, css change for padding
If you include values for two or three properties, the CSS padding order of values changes: One value — sets the same padding to all four sides. Two values — the first value applies the padding on top and bottom. The second adds padding to the left and right sides of an element.